add avatarimage to toast
