+++ backend/src/modules/notification/index.ts
@@ -1,7 +1,7 @@
 import { Notiobject } from "@apiType/notification"
 import express from "express"
 import { Server as IOServer, Socket } from "socket.io"
-import getUserNotiObject from "./routes/getUserNotiObject"
+
 import markallasRead from "./routes/markallasRead"
 import readNotiObject from "./routes/readNotiObject"
 import getNotiUser from "./routes/getNotiUser"
@@ -24,7 +24,6 @@ notificationRoutes.use((req, res, next) => {
     next()
 })
 
-notificationRoutes.get("/getusernotiobject", getUserNotiObject)
 notificationRoutes.get("/getNotiUser", getNotiUser)
 notificationRoutes.get("/getusernotiobjectbymodule/:module", getUserNotiObjectbyModule)
 notificationRoutes.get("/getsenderimage/:senderId", getSenderImage)
